Added file/folder existence checks, restructured

+++ burn-me.sh
@@ -1,16 +1,6 @@
 #!/bin/bash
 
-echo -e "\n
-    #################################################### 
-    WARNING! BURN NOTICE WILL IRREVERSIBLY DESTROY DATA! 
-    #################################################### 
-    \n\nPress <ENTER> to continue ONLY if you are sure this is what you want. Otherwise, quit this script."
-
-read -r
-
-echo "Burning..."
-
-# Remove hidden files that may be sensitive
+# List of hidden files to burn
 LISTFILES="
     .aws 
     .bash_history
@@ -28,11 +18,7 @@ LISTFILES="
     .zshrc
 "
 
-for ITEM in $LISTFILES ; do
-    rm -rf "$HOME"/"${ITEM:?}"
-done
-
-# Remove the contents of the user's most used directories
+# List of directories to burn
 LISTDIRECTORIES="
     Desktop/*
     Documents/*
@@ -40,10 +26,40 @@ LISTDIRECTORIES="
     Movies/*
     Music/*
     Pictures/*
+    git/*
 "
 
-for ITEM in $LISTDIRECTORIES ; do
-    cd "$HOME"/"$ITEM" || exit
-    rm -rf .[^.]*
-    cd .. || exit
+# Run the script
+echo -e "\n
+    #################################################### 
+    WARNING! BURN NOTICE WILL IRREVERSIBLY DESTROY DATA! 
+    #################################################### 
+    \n\nPress <ENTER> to continue ONLY if you are sure this is what you want. Otherwise, quit this script."
+
+read -r
+
+echo "Burning..."
+
+# Remove hidden files that may be sensitive
+for ITEM in $LISTFILES ; do
+    if [ -f "$HOME"/"$ITEM" ] ; then
+        rm -rf "$HOME"/"${ITEM:?}"
+    else
+        echo "$HOME/$ITEM doesn't exist, skipping."
+    fi
 done
+
+# Remove the contents of the user's most used directories
+for ITEM in $LISTDIRECTORIES ; do
+    if [ -f "$HOME"/"$ITEM" ] ; then
+        cd "$HOME"/"$ITEM" || exit
+        rm -rf .[^.]*
+        cd .. || exit
+    else
+        echo "$HOME/$ITEM doesn't exist, skipping."
+    fi
+done
+
+echo -e "\nThe script is complete. To fully burn your identity, it's suggested to also remove browsing data, log out of iCloud and other installed apps, and if very paranoid, format your hard drive."
+
+history -c
